refactor(api): extract content ops into content_ops.py; reorder tags
operations.py had grown to 1519 lines mixing two distinct surfaces. Split the
seven content operations (remember/forget/update-memory/index-document/
delete-document/ingest-file/delete-file) and their five content-op-only guards
into routers/graphs/content_ops.py; operations.py keeps the graph-lifecycle ops
(subgraph/backup/tier/style/materialize) and the shared envelope-dispatch
machinery (_ctx/_dispatch/constants), which content_ops imports one-way.

Pure move — handlers, routes, tags, and behavior are unchanged; both routers
mount at /operations. Also move the "Graph Operations" tag up next to "Graphs"
in the OpenAPI tag order (it's a core graph surface, not an afterthought).

Verified: routes mount correctly (7 content / 8 lifecycle); 579 router tests +
test-code green; live smoke — index-document 200 via content_ops, materialize
202 via operations, tags correct.
